Move hooks into module_init dir
