+++ app/src/main/java/com/qnotifiedx/app/hook/base/moduleinit/GetAppContext.kt
@@ -1,22 +1,18 @@
-package com.qnotifiedx.app.hook.normal
+package com.qnotifiedx.app.hook.base.moduleinit
 
 import android.app.Application
-import com.qnotifiedx.annotations.NormalHookEntry
-import com.qnotifiedx.app.hook.base.BaseNormalHook
 import com.qnotifiedx.app.util.*
 
 //获取宿主全局ApplicationHook
-@NormalHookEntry
-object GetAppContext : BaseNormalHook() {
-    //强制开启
-    override var enable: Boolean = true
+
+object GetAppContext {
     var application: Application? = null
         private set
 
-    override fun init() {
+    fun init() {
         for (m in getMethods("com.tencent.mobileqq.startup.step.LoadDex")) {
             if (m.returnType == Boolean::class.javaPrimitiveType && m.parameterTypes.isEmpty()) {
-                m.hookAfter {
+                m.hookAfter(100) {
                     if (application != null) return@hookAfter
                     //加载QQ的基础Application
                     val cBaseApplicationImpl =
