+++ app/src/main/java/com/qnotifiedx/app/hook/base/moduleinit/ModuleEntry.kt
@@ -1,24 +1,19 @@
-package com.qnotifiedx.app.hook.normal
+package com.qnotifiedx.app.hook.base.moduleinit
 
 import android.content.Context
 import android.view.View
 import android.view.ViewGroup
 import androidx.core.view.size
-import com.qnotifiedx.annotations.NormalHookEntry
 import com.qnotifiedx.app.BuildConfig
-import com.qnotifiedx.app.hook.base.BaseNormalHook
 import com.qnotifiedx.app.util.*
 
 //模块入口Hook
-@NormalHookEntry
-object ModuleEntry : BaseNormalHook() {
+object ModuleEntry {
     //强制开启
-    override var enable: Boolean = true
-
-    override fun init() {
+    fun init() {
         for (m in getMethods("com.tencent.mobileqq.activity.QQSettingSettingActivity")) {
             if (m.name != "doOnCreate") continue
-            m.hookAfter {
+            m.hookAfter(100) {
                 val thisObject = it.thisObject
                 //加载QQ的设置物件类
                 val cFormSimpleItem = loadClass("com.tencent.mobileqq.widget.FormSimpleItem")
